+++ apps/decodex/src/radar.rs
@@ -25,8 +25,10 @@ mod github_bundle_client;
 mod github_token;
 mod ledger;
 mod requests;
+mod review_queue;
 mod signal_render;
 mod social_publish;
+mod source_bundle;
 
 #[cfg(test)]
 use artifact_validation::has_legacy_multi_agent_v2_context;
@@ -40,8 +42,11 @@ use github_api::GitHubApi;
 use github_bundle_client::GithubClient;
 use github_token::github_token;
 use ledger::RadarLedger;
+use review_queue::RecentCommit;
+use review_queue::build_review_queue;
 use signal_render::{rendered_config_flags, rendered_signal};
 pub(crate) use social_publish::reserve_social_publish;
+use source_bundle::{build_commit_bundle_from_sources, build_pr_bundle_from_sources};
 
 pub(crate) use ledger::{
 	default_ledger_path, ledger_artifact_link, ledger_bootstrap, ledger_ingest,
@@ -207,47 +212,6 @@ struct BackfillPaths {
 	bundle: PathBuf,
 	analysis: PathBuf,
 	signal: PathBuf,
-}
-
-struct RecentCommit {
-	sha: String,
-	title: String,
-	url: String,
-	committed_at: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-struct BundleFile {
-	path: String,
-	patch_excerpt: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-struct BundleCommit {
-	sha: String,
-	message: String,
-}
-
-#[derive(Clone, Debug)]
-struct BundlePr {
-	number: u64,
-	title: String,
-	body: String,
-	state: String,
-	url: String,
-}
-
-#[derive(Clone, Debug)]
-struct SourceBundle {
-	primary_pr: Option<BundlePr>,
-	commits: Vec<BundleCommit>,
-	files: Vec<BundleFile>,
-}
-
-#[derive(Debug)]
-struct QueueBuild {
-	queue: Value,
-	ledger_enabled: bool,
 }
 
 #[derive(Clone, Debug)]
@@ -521,91 +485,6 @@ fn validate_expected_schema(
 	Ok(())
 }
 
-fn build_review_queue(
-	request: &RadarRefreshQueueRequest,
-	root: &Path,
-	api: &GitHubApi,
-) -> crate::prelude::Result<QueueBuild> {
-	let (default_branch, commits) = recent_commits(api, &request.repo, request.search_limit)?;
-	let recent_commits_scanned = commits.len();
-	let (published_prs, published_shas) =
-		published_subjects(&absolute_repo_path(root, &request.signals_dir))?;
-	let ledger_path = ledger_path(root, request);
-	let mut ledger = ledger_path.as_deref().map(RadarLedger::open).transpose()?;
-	let mut subjects = BTreeMap::<(String, String), Value>::new();
-	let mut published_seen = 0_usize;
-
-	for commit in commits {
-		let pr_number = maybe_promote_commit_to_pr(api, &request.repo, &commit.sha)?;
-		let subject_kind = if pr_number.is_some() { "pr" } else { "commit" };
-		let subject_id = pr_number.map_or_else(|| commit.sha.clone(), |number| number.to_string());
-
-		if let Some(ledger) = &mut ledger {
-			ledger.record_commit(&request.repo, &commit, pr_number)?;
-		}
-
-		if published_shas.contains(&commit.sha)
-			|| pr_number.is_some_and(|number| published_prs.contains(&number))
-		{
-			published_seen += 1;
-
-			if let Some(ledger) = &mut ledger {
-				ledger.record_review(
-					&request.repo,
-					subject_kind,
-					&subject_id,
-					"signal",
-					"Already present in published signal collection.",
-					Some("confirmed"),
-				)?;
-			}
-
-			continue;
-		}
-
-		let key = (subject_kind.to_owned(), subject_id.clone());
-
-		if let Some(current) = subjects.get_mut(&key) {
-			append_commit_sha(current, &commit.sha);
-
-			continue;
-		}
-
-		let bundle = match pr_number {
-			Some(number) => build_pr_bundle(api, &request.repo, number)?,
-			None => build_commit_bundle(api, &request.repo, &commit.sha)?,
-		};
-
-		subjects.insert(key, subject_from_bundle(&bundle, subject_kind, &subject_id, &commit));
-
-		if let Some(ledger) = &mut ledger {
-			ledger.record_review(
-				&request.repo,
-				subject_kind,
-				&subject_id,
-				"watch",
-				"Queued for AI upstream review by deterministic Radar sync.",
-				Some("likely"),
-			)?;
-		}
-	}
-
-	if let Some(ledger) = &mut ledger {
-		ledger.commit()?;
-	}
-
-	let ordered_subjects = sort_queue_subjects(subjects.into_values().collect());
-	let queue = review_queue_payload(
-		request,
-		&default_branch,
-		recent_commits_scanned,
-		published_seen,
-		ordered_subjects,
-	)?;
-
-	Ok(QueueBuild { queue, ledger_enabled: !request.no_ledger })
-}
-
 fn build_release_delta(
 	request: &RadarRefreshReleaseDeltaRequest,
 	root: &Path,
@@ -672,358 +551,6 @@ fn build_release_delta(
 		"comparisons": comparison_entries,
 		"tracked_signal_slugs": default_tracked_signal_slugs,
 	}))
-}
-
-fn review_queue_payload(
-	request: &RadarRefreshQueueRequest,
-	default_branch: &str,
-	recent_commits_scanned: usize,
-	published_seen: usize,
-	subjects: Vec<Value>,
-) -> crate::prelude::Result<Value> {
-	let critical = count_priority(&subjects, "critical");
-	let high = count_priority(&subjects, "high");
-	let normal = count_priority(&subjects, "normal");
-	let low = count_priority(&subjects, "low");
-
-	Ok(serde_json::json!({
-		"schema": UPSTREAM_REVIEW_QUEUE_SCHEMA,
-		"repo": request.repo,
-		"generated_at": utc_now_iso()?,
-		"source": {
-			"default_branch": default_branch,
-			"search_limit": request.search_limit,
-			"signals_dir": request.signals_dir.to_string_lossy(),
-		},
-		"subjects": subjects,
-		"counts": {
-			"recent_commits_scanned": recent_commits_scanned,
-			"published_subjects_seen": published_seen,
-			"subjects_queued": critical + high + normal + low,
-			"critical": critical,
-			"high": high,
-			"normal": normal,
-			"low": low,
-		},
-	}))
-}
-
-fn count_priority(subjects: &[Value], priority: &str) -> usize {
-	subjects
-		.iter()
-		.filter(|subject| {
-			subject
-				.get("review_priority")
-				.and_then(Value::as_str)
-				.is_some_and(|value| value == priority)
-		})
-		.count()
-}
-
-fn recent_commits(
-	api: &GitHubApi,
-	repo: &str,
-	search_limit: usize,
-) -> crate::prelude::Result<(String, Vec<RecentCommit>)> {
-	let default_branch = repo_default_branch(api, repo)?;
-	let url = format!(
-		"https://api.github.com/repos/{repo}/commits?sha={}&per_page={search_limit}",
-		percent_encode(&default_branch)
-	);
-	let payload = api.get(&url)?.payload;
-	let Some(items) = payload.as_array() else {
-		eyre::bail!("Expected commits list payload from GitHub API");
-	};
-	let commits = items.iter().filter_map(recent_commit_from_value).collect::<Vec<_>>();
-
-	Ok((default_branch, commits))
-}
-
-fn recent_commit_from_value(item: &Value) -> Option<RecentCommit> {
-	let commit = item.get("commit")?.as_object()?;
-	let sha = item.get("sha")?.as_str()?.to_owned();
-	let url = item.get("html_url")?.as_str()?.to_owned();
-	let message = commit.get("message")?.as_str()?;
-
-	if message.is_empty() {
-		return None;
-	}
-
-	Some(RecentCommit {
-		sha,
-		title: first_line(message),
-		url,
-		committed_at: commit
-			.get("committer")
-			.and_then(Value::as_object)
-			.and_then(|committer| committer.get("date"))
-			.and_then(Value::as_str)
-			.map(str::to_owned),
-	})
-}
-
-fn published_subjects(
-	signals_dir: &Path,
-) -> crate::prelude::Result<(HashSet<u64>, HashSet<String>)> {
-	let mut published_prs = HashSet::new();
-	let mut published_shas = HashSet::new();
-
-	for path in sorted_json_files(signals_dir)? {
-		let payload = load_json(&path)?;
-
-		validate_signal_file(&path, &payload)?;
-
-		if let Some(pr_number) = payload
-			.get("source_refs")
-			.and_then(|refs| refs.get("pr_url"))
-			.and_then(Value::as_str)
-			.and_then(extract_pr_number_from_url)
-		{
-			published_prs.insert(pr_number);
-		}
-
-		for url in string_array(payload.pointer("/source_refs/commit_urls")) {
-			if let Some(sha) = extract_commit_sha_from_url(&url) {
-				published_shas.insert(sha);
-			}
-		}
-	}
-
-	Ok((published_prs, published_shas))
-}
-
-fn maybe_promote_commit_to_pr(
-	api: &GitHubApi,
-	repo: &str,
-	commit_sha: &str,
-) -> crate::prelude::Result<Option<u64>> {
-	let url = format!("https://api.github.com/repos/{repo}/commits/{commit_sha}/pulls");
-	let pulls = match api.get_paginated(&url) {
-		Ok(pulls) => pulls,
-		Err(_) => return Ok(None),
-	};
-
-	Ok(pulls.first().and_then(|first| first.get("number")).and_then(Value::as_u64))
-}
-
-fn build_pr_bundle(
-	api: &GitHubApi,
-	repo: &str,
-	pr_number: u64,
-) -> crate::prelude::Result<SourceBundle> {
-	let pr = api.get(&format!("https://api.github.com/repos/{repo}/pulls/{pr_number}"))?.payload;
-	let commits = api.get_paginated(&format!(
-		"https://api.github.com/repos/{repo}/pulls/{pr_number}/commits?per_page=100"
-	))?;
-	let files = api.get_paginated(&format!(
-		"https://api.github.com/repos/{repo}/pulls/{pr_number}/files?per_page=100"
-	))?;
-
-	Ok(SourceBundle {
-		primary_pr: Some(BundlePr {
-			number: required_value_u64(&pr, "number")?,
-			title: required_value_string(&pr, "title")?,
-			body: optional_value_string(&pr, "body").unwrap_or_default(),
-			state: if optional_value_string(&pr, "merged_at").is_some() {
-				"merged".to_owned()
-			} else {
-				required_value_string(&pr, "state")?
-			},
-			url: required_value_string(&pr, "html_url")?,
-		}),
-		commits: commits.iter().filter_map(bundle_commit_from_pr_commit).collect(),
-		files: files.iter().filter_map(bundle_file_from_value).collect(),
-	})
-}
-
-fn build_commit_bundle(
-	api: &GitHubApi,
-	repo: &str,
-	commit_sha: &str,
-) -> crate::prelude::Result<SourceBundle> {
-	let commit =
-		api.get(&format!("https://api.github.com/repos/{repo}/commits/{commit_sha}"))?.payload;
-	let files = commit.get("files").and_then(Value::as_array).cloned().unwrap_or_default();
-	let message = commit.pointer("/commit/message").and_then(Value::as_str).unwrap_or_default();
-
-	Ok(SourceBundle {
-		primary_pr: None,
-		commits: vec![BundleCommit {
-			sha: required_value_string(&commit, "sha")?,
-			message: first_line(message),
-		}],
-		files: files.iter().filter_map(bundle_file_from_value).collect(),
-	})
-}
-
-fn bundle_commit_from_pr_commit(item: &Value) -> Option<BundleCommit> {
-	Some(BundleCommit {
-		sha: item.get("sha")?.as_str()?.to_owned(),
-		message: first_line(item.pointer("/commit/message")?.as_str()?),
-	})
-}
-
-fn bundle_file_from_value(item: &Value) -> Option<BundleFile> {
-	Some(BundleFile {
-		path: item.get("filename")?.as_str()?.to_owned(),
-		patch_excerpt: item.get("patch").and_then(Value::as_str).map(truncate_patch_excerpt),
-	})
-}
-
-fn subject_from_bundle(
-	bundle: &SourceBundle,
-	subject_kind: &str,
-	subject_id: &str,
-	seed_commit: &RecentCommit,
-) -> Value {
-	let surface_hints = detect_surface_hints(bundle);
-	let attention_flags = detect_attention_flags(bundle);
-	let mut subject = serde_json::json!({
-		"subject_kind": subject_kind,
-		"subject_id": subject_id,
-		"title": seed_commit.title.clone(),
-		"url": seed_commit.url.clone(),
-		"source_state": "commit_only",
-		"commit_shas": commit_shas(bundle, seed_commit),
-		"committed_at": seed_commit.committed_at.clone(),
-		"changed_file_count": bundle.files.len(),
-		"sample_paths": bundle.files.iter().take(12).map(|file| file.path.clone()).collect::<Vec<_>>(),
-		"surface_hints": surface_hints,
-		"attention_flags": attention_flags,
-		"review_priority": priority_for(&surface_hints, &attention_flags),
-		"review_reason": review_reason(&surface_hints, &attention_flags),
-		"next_step": "ai_review_required",
-	});
-
-	if let Some(primary_pr) = &bundle.primary_pr
-		&& let Some(subject) = subject.as_object_mut()
-	{
-		subject.insert("title".to_owned(), Value::String(primary_pr.title.clone()));
-		subject.insert("url".to_owned(), Value::String(primary_pr.url.clone()));
-		subject.insert("source_state".to_owned(), Value::String(primary_pr.state.clone()));
-		subject.insert("pr_number".to_owned(), Value::from(primary_pr.number));
-		subject.insert("pr_url".to_owned(), Value::String(primary_pr.url.clone()));
-	}
-
-	subject
-}
-
-fn commit_shas(bundle: &SourceBundle, seed_commit: &RecentCommit) -> Vec<String> {
-	let shas = bundle.commits.iter().map(|commit| commit.sha.clone()).collect::<Vec<_>>();
-
-	if shas.is_empty() { vec![seed_commit.sha.clone()] } else { shas }
-}
-
-fn append_commit_sha(subject: &mut Value, sha: &str) {
-	let Some(shas) = subject.get_mut("commit_shas").and_then(Value::as_array_mut) else {
-		return;
-	};
-
-	if !shas.iter().any(|value| value.as_str() == Some(sha)) {
-		shas.push(Value::String(sha.to_owned()));
-	}
-}
-
-fn sort_queue_subjects(mut subjects: Vec<Value>) -> Vec<Value> {
-	subjects.sort_by_key(queue_sort_key);
-
-	subjects
-}
-
-fn queue_sort_key(subject: &Value) -> (u8, String, String, String) {
-	(
-		match subject.get("review_priority").and_then(Value::as_str) {
-			Some("critical") => 0,
-			Some("high") => 1,
-			Some("normal") => 2,
-			Some("low") => 3,
-			_ => 9,
-		},
-		subject.get("committed_at").and_then(Value::as_str).unwrap_or_default().to_owned(),
-		subject.get("subject_kind").and_then(Value::as_str).unwrap_or_default().to_owned(),
-		subject.get("subject_id").and_then(Value::as_str).unwrap_or_default().to_owned(),
-	)
-}
-
-fn detect_surface_hints(bundle: &SourceBundle) -> Vec<String> {
-	let haystack =
-		bundle.files.iter().map(|file| file.path.to_lowercase()).collect::<Vec<_>>().join("\n");
-	let mut hints = SURFACE_RULES
-		.iter()
-		.filter(|(_, terms)| terms.iter().any(|term| haystack.contains(term)))
-		.map(|(surface, _)| (*surface).to_owned())
-		.collect::<Vec<_>>();
-
-	if hints.is_empty() {
-		hints.push("internal_churn".to_owned());
-	}
-
-	hints.sort();
-
-	hints
-}
-
-fn detect_attention_flags(bundle: &SourceBundle) -> Vec<String> {
-	let haystack = text_blob(bundle);
-	let mut flags = ATTENTION_RULES
-		.iter()
-		.filter(|(_, terms)| terms.iter().any(|term| haystack.contains(term)))
-		.map(|(flag, _)| (*flag).to_owned())
-		.collect::<Vec<_>>();
-
-	flags.sort();
-
-	flags
-}
-
-fn text_blob(bundle: &SourceBundle) -> String {
-	let mut parts = Vec::new();
-
-	if let Some(primary_pr) = &bundle.primary_pr {
-		parts.push(primary_pr.title.clone());
-		parts.push(primary_pr.body.clone());
-	}
-
-	parts.extend(bundle.commits.iter().map(|commit| commit.message.clone()));
-	parts.extend(
-		bundle
-			.files
-			.iter()
-			.flat_map(|file| [file.path.clone(), file.patch_excerpt.clone().unwrap_or_default()]),
-	);
-
-	parts.join("\n").to_lowercase()
-}
-
-fn priority_for(surface_hints: &[String], attention_flags: &[String]) -> &'static str {
-	let has_high_surface =
-		surface_hints.iter().any(|surface| HIGH_VALUE_SURFACES.contains(&surface.as_str()));
-	let breaking_or_removed = attention_flags
-		.iter()
-		.any(|flag| matches!(flag.as_str(), "breaking_change" | "deprecated_removed"));
-
-	if breaking_or_removed && has_high_surface {
-		"critical"
-	} else if has_high_surface {
-		"high"
-	} else if attention_flags.iter().any(|flag| {
-		matches!(flag.as_str(), "new_feature" | "protocol_change" | "release_packaging")
-	}) {
-		"normal"
-	} else {
-		"low"
-	}
-}
-
-fn review_reason(surface_hints: &[String], attention_flags: &[String]) -> String {
-	if surface_hints.iter().any(|hint| hint == "internal_churn") && attention_flags.is_empty() {
-		return "Needs AI review because every recent upstream commit is tracked, but deterministic hints found only internal churn.".to_owned();
-	}
-	if !attention_flags.is_empty() {
-		return format!("Needs AI review for {}.", attention_flags.join(", "));
-	}
-
-	format!("Needs AI review for surface hints: {}.", surface_hints.join(", "))
 }
 
 fn github_releases(api: &GitHubApi, repo: &str) -> crate::prelude::Result<Vec<Value>> {
@@ -2166,312 +1693,6 @@ fn is_truthy_json_value(value: Option<&Value>) -> bool {
 
 fn non_empty_array(value: Option<&Value>) -> Option<&Vec<Value>> {
 	value.and_then(Value::as_array).filter(|values| !values.is_empty())
-}
-
-fn build_pr_bundle_from_sources(
-	repo: &str,
-	pr: &Value,
-	commits: &[Value],
-	files: &[Value],
-	default_branch: &str,
-	notes: &[String],
-) -> crate::prelude::Result<Value> {
-	let pr = object_value(pr, "pull request")?;
-	let commit_items =
-		commits.iter().map(commit_bundle_item).collect::<crate::prelude::Result<Vec<_>>>()?;
-	let file_items =
-		files.iter().map(file_bundle_item).collect::<crate::prelude::Result<Vec<_>>>()?;
-	let docs_refs = collect_docs_refs(files);
-	let examples_refs = collect_examples_refs(files);
-	let all_patch_text = files
-		.iter()
-		.filter_map(|file| file.get("patch").and_then(Value::as_str))
-		.collect::<Vec<_>>()
-		.join("\n");
-	let all_commit_text = commits
-		.iter()
-		.filter_map(|commit| {
-			commit
-				.get("commit")
-				.and_then(Value::as_object)
-				.and_then(|commit| commit.get("message"))
-				.and_then(Value::as_str)
-		})
-		.collect::<Vec<_>>()
-		.join("\n");
-	let mut bundle_notes =
-		vec!["Built from GitHub pull-request, commits, files, and repo endpoints.".to_owned()];
-
-	bundle_notes.extend(notes.iter().cloned());
-
-	let primary_pr = serde_json::json!({
-		"number": required_u64(pr, "number", "primary_pr.number")?,
-		"title": required_string(pr, "title", "primary_pr.title")?,
-		"body": pr.get("body").and_then(Value::as_str).unwrap_or(""),
-		"state": pr
-			.get("merged_at")
-			.and_then(Value::as_str)
-			.filter(|value| !value.is_empty())
-			.map_or_else(
-				|| required_string(pr, "state", "primary_pr.state").map(str::to_owned),
-				|_| Ok("merged".to_owned()),
-			)?,
-		"merged_at": pr.get("merged_at").cloned().unwrap_or(Value::Null),
-		"labels": pr_labels(pr),
-		"url": required_string(pr, "html_url", "primary_pr.url")?,
-	});
-	let bundle = serde_json::json!({
-		"schema": BUNDLE_SCHEMA,
-		"repo": repo,
-		"analysis_mode": "pr_first",
-		"default_branch": default_branch,
-		"primary_pr": primary_pr,
-		"commits": commit_items,
-		"files": file_items,
-		"linked_issues": collect_issue_refs(
-			&[pr.get("body").and_then(Value::as_str).unwrap_or(""), &all_commit_text]
-		)?,
-		"extracted_flags": collect_flags(&[
-			pr.get("body").and_then(Value::as_str).unwrap_or(""),
-			&all_commit_text,
-			&all_patch_text,
-		])?,
-		"docs_refs": docs_refs,
-		"examples_refs": examples_refs,
-		"notes": bundle_notes,
-	});
-
-	validate_bundle_value(&bundle)?;
-
-	Ok(bundle)
-}
-
-fn build_commit_bundle_from_sources(
-	repo: &str,
-	commit: &Value,
-	default_branch: &str,
-	notes: &[String],
-) -> crate::prelude::Result<Value> {
-	let commit = object_value(commit, "commit")?;
-	let files = commit.get("files").and_then(Value::as_array).cloned().unwrap_or_default();
-	let commit_payload = object_field(commit, "commit", "commit.commit")?;
-	let commit_message = required_string(commit_payload, "message", "commit.commit.message")?;
-	let all_patch_text = files
-		.iter()
-		.filter_map(|file| file.get("patch").and_then(Value::as_str))
-		.collect::<Vec<_>>()
-		.join("\n");
-	let mut bundle_notes = vec!["Built from GitHub commit endpoint without PR context.".to_owned()];
-
-	bundle_notes.extend(notes.iter().cloned());
-
-	let bundle = serde_json::json!({
-		"schema": BUNDLE_SCHEMA,
-		"repo": repo,
-		"analysis_mode": "commit_only",
-		"default_branch": default_branch,
-		"commits": [commit_bundle_item(&Value::Object(commit.clone()))?],
-		"files": files
-			.iter()
-			.map(file_bundle_item)
-			.collect::<crate::prelude::Result<Vec<_>>>()?,
-		"linked_issues": collect_issue_refs(&[commit_message])?,
-		"extracted_flags": collect_flags(&[commit_message, &all_patch_text])?,
-		"docs_refs": collect_docs_refs(&files),
-		"examples_refs": collect_examples_refs(&files),
-		"notes": bundle_notes,
-	});
-
-	validate_bundle_value(&bundle)?;
-
-	Ok(bundle)
-}
-
-fn commit_bundle_item(commit: &Value) -> crate::prelude::Result<Value> {
-	let commit = object_value(commit, "commit")?;
-	let payload = object_field(commit, "commit", "commit.commit")?;
-	let author = object_field(payload, "author", "commit.commit.author").ok();
-	let author_name = commit
-		.get("author")
-		.and_then(Value::as_object)
-		.and_then(|author| author.get("login"))
-		.and_then(Value::as_str)
-		.or_else(|| author.and_then(|author| author.get("name")).and_then(Value::as_str));
-	let committed_at = author.and_then(|author| author.get("date")).cloned().unwrap_or(Value::Null);
-
-	Ok(serde_json::json!({
-		"sha": required_string(commit, "sha", "commit.sha")?,
-		"message": first_line(required_string(payload, "message", "commit.commit.message")?),
-		"url": required_string(commit, "html_url", "commit.html_url")?,
-		"author": author_name,
-		"committed_at": committed_at,
-	}))
-}
-
-fn file_bundle_item(file: &Value) -> crate::prelude::Result<Value> {
-	let file = object_value(file, "file")?;
-
-	Ok(serde_json::json!({
-		"path": required_string(file, "filename", "file.filename")?,
-		"status": required_string(file, "status", "file.status")?,
-		"additions": required_i64(file, "additions", "file.additions")?,
-		"deletions": required_i64(file, "deletions", "file.deletions")?,
-		"patch_excerpt": file
-			.get("patch")
-			.and_then(Value::as_str)
-			.and_then(truncate_patch),
-	}))
-}
-
-fn validate_bundle_value(bundle: &Value) -> crate::prelude::Result<()> {
-	let validation = validate_artifact(bundle);
-
-	if validation.errors.is_empty() && validation.schema.as_deref() == Some(BUNDLE_SCHEMA) {
-		Ok(())
-	} else {
-		let mut errors = validation.errors;
-
-		if validation.schema.as_deref() != Some(BUNDLE_SCHEMA) {
-			errors.insert(0, format!("schema must be {BUNDLE_SCHEMA}"));
-		}
-
-		eyre::bail!("Bundle validation failed:\n- {}", errors.join("\n- "))
-	}
-}
-
-fn object_field<'a>(
-	object: &'a Map<String, Value>,
-	field: &str,
-	label: &str,
-) -> crate::prelude::Result<&'a Map<String, Value>> {
-	object
-		.get(field)
-		.and_then(Value::as_object)
-		.ok_or_else(|| eyre::eyre!("{label} must be an object"))
-}
-
-fn required_u64(
-	object: &Map<String, Value>,
-	field: &str,
-	label: &str,
-) -> crate::prelude::Result<u64> {
-	object
-		.get(field)
-		.and_then(Value::as_u64)
-		.ok_or_else(|| eyre::eyre!("{label} must be an unsigned integer"))
-}
-
-fn required_i64(
-	object: &Map<String, Value>,
-	field: &str,
-	label: &str,
-) -> crate::prelude::Result<i64> {
-	object
-		.get(field)
-		.and_then(Value::as_i64)
-		.ok_or_else(|| eyre::eyre!("{label} must be an integer"))
-}
-
-fn pr_labels(pr: &Map<String, Value>) -> Vec<String> {
-	pr.get("labels")
-		.and_then(Value::as_array)
-		.into_iter()
-		.flatten()
-		.filter_map(|label| {
-			label
-				.as_object()
-				.and_then(|label| label.get("name"))
-				.and_then(Value::as_str)
-				.map(str::to_owned)
-		})
-		.collect()
-}
-
-fn collect_docs_refs(files: &[Value]) -> Vec<String> {
-	files
-		.iter()
-		.filter_map(file_name)
-		.filter(|filename| filename.starts_with("docs/") || filename.ends_with("README.md"))
-		.map(str::to_owned)
-		.collect()
-}
-
-fn collect_examples_refs(files: &[Value]) -> Vec<String> {
-	files
-		.iter()
-		.filter_map(file_name)
-		.filter(|filename| {
-			filename.to_lowercase().contains("example") || filename.contains("examples/")
-		})
-		.map(str::to_owned)
-		.collect()
-}
-
-fn file_name(file: &Value) -> Option<&str> {
-	file.as_object()?.get("filename")?.as_str()
-}
-
-fn collect_issue_refs(texts: &[&str]) -> crate::prelude::Result<Vec<String>> {
-	collect_regex_matches(issue_ref_regex()?, texts)
-}
-
-fn collect_flags(texts: &[&str]) -> crate::prelude::Result<Vec<String>> {
-	collect_regex_matches(flag_regex()?, texts)
-}
-
-fn collect_regex_matches(regex: &Regex, texts: &[&str]) -> crate::prelude::Result<Vec<String>> {
-	let mut found = Vec::new();
-
-	for text in texts {
-		for captures in regex.captures_iter(text) {
-			let Some(value) = captures.get(1).map(|matched| matched.as_str()) else {
-				continue;
-			};
-
-			if !found.iter().any(|found_value| found_value == value) {
-				found.push(value.to_owned());
-			}
-		}
-	}
-
-	Ok(found)
-}
-
-fn issue_ref_regex() -> crate::prelude::Result<&'static Regex> {
-	static ISSUE_REF_RE: OnceLock<std::result::Result<Regex, regex::Error>> = OnceLock::new();
-
-	ISSUE_REF_RE
-		.get_or_init(|| Regex::new(r"(?:^|[^\w])((?:[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+)?#\d+)"))
-		.as_ref()
-		.map_err(|error| eyre::eyre!("Failed to compile issue reference regex: {error}"))
-}
-
-fn flag_regex() -> crate::prelude::Result<&'static Regex> {
-	static FLAG_RE: OnceLock<std::result::Result<Regex, regex::Error>> = OnceLock::new();
-
-	FLAG_RE
-		.get_or_init(|| {
-			Regex::new(r"(?:^|[^\w-])(--[a-zA-Z0-9][\w-]*|[A-Z][A-Z0-9_]{2,}(?:=[^\s,`]+)?)")
-		})
-		.as_ref()
-		.map_err(|error| eyre::eyre!("Failed to compile flag regex: {error}"))
-}
-
-fn truncate_patch(value: &str) -> Option<String> {
-	let compact = value.trim();
-
-	if compact.is_empty() {
-		return None;
-	}
-	if compact.chars().count() > 900 {
-		let mut truncated = compact.chars().take(900).collect::<String>();
-
-		truncated.push_str("...");
-
-		Some(truncated)
-	} else {
-		Some(compact.into())
-	}
 }
 
 fn first_line(value: &str) -> String {
